Create api.js
Adding the api.js file to the project's repo. I've written all the code logic for the app.route('/api/threads/:board'). Posting, reporting and deleting a thread to and from a specific board works as intended. The code logic for the app.route('/api/replies/:board') is all that remains to be written to complete the project. As per FCC recommendation the controllers/handlers and handle routing will be done in routes/api.js, so no separate modules for this project.