Update board-style.css
Adding the additional board-style.css file to the project's repo. I decided to separate the styling for the board.html and thread.html from the one for the index.html for better styling management. The code probably needs to be optimized but that will be done in a future update.
Update_07.08.2019:  The final mods to this file.