Update main-style.css
Adding the main-style.css file to the project's repo. Being based on my previous projects from the series of ISQA Projects, the css file contains a few bits of additional styling that will be cleaned in a future update.

Update_07.08.2019:  Cleaned the file of the old styling. The final mods to this file.