Update api.js
Adding the api.js file to the project's repo. I've written all the code logic for the app.route('/api/threads/:board'). Posting, reporting and deleting a thread to and from a specific board works as intended. The code logic for the app.route('/api/replies/:board') is all that remains to be written to complete the project. As per FCC recommendation the controllers/handlers and handle routing will be done in routes/api.js, so no separate modules for this project.

Update_07.08.2019: Written all the methods for the replies route and optimized the code overall. The project is now complete, final and fully functional. Thoroughly tested it manually and completed the functional_tests. The project is ready to be submitted to FCC.